requirements.txt
# ============================================
# GenAI & Image Generation
# ============================================
openai>=1.0.0              # OpenAI API for DALL-E 3 image generation
google-generativeai>=0.8.0 # Google Gemini API (required for agent module)

# ============================================
# HTTP & API Requests
# ============================================
requests>=2.31.0           # HTTP library for API calls

# ============================================
# YAML Config Support
# ============================================
pyyaml>=6.0                # YAML parser for config file

# ============================================
# Image Processing
# ============================================
pillow>=10.0.0             # PIL for image manipulation and resizing
opencv-python>=4.8.0       # OpenCV for logo compliance checking (feature detection)

# ============================================
# Cloud Storage (Optional)
# ============================================
dropbox>=11.36.0           # Dropbox SDK for file upload/download

# ============================================
# Agent Module (File Watching)
# ============================================
watchdog>=3.0.0            # File system event monitoring for agent watcher
